typing and moving to analysisAPI
