+++ cg/meta/workflow/raredisease.py
@@ -2,9 +2,8 @@
 
 import logging
 from pathlib import Path
-from typing import Any
 
-from cg.constants import DEFAULT_CAPTURE_KIT, Workflow
+from cg.constants import Workflow
 from cg.constants.constants import AnalysisType, GenomeVersion
 from cg.constants.gene_panel import GenePanelGenomeBuild
 from cg.constants.nf_analysis import RAREDISEASE_METRIC_CONDITIONS
@@ -74,34 +73,10 @@ class RarediseaseAnalysisAPI(NfAnalysisAPI):
         )
         return sample_sheet_entry.reformat_sample_content
 
-    def get_analysis_type(self, case_id: str):
-        case = self.get_case_from_string(case_id)
-        sample_analysis_type = ""
-        for link in case.links:
-            case_sample = link
-            sample_analysis_type_tmp = (
-                case_sample.sample.application_version.application.analysis_type
-            )
-            if sample_analysis_type_tmp != sample_analysis_type and sample_analysis_type != "":
-                raise ValueError(
-                    f"{sample_analysis_type_tmp} has not the same analysis type as other samples in the case"
-                )
-            sample_analysis_type = sample_analysis_type_tmp
-        return sample_analysis_type
-
-    def set_target_bed(self, case_id: str, analysis_type: str) -> str:
-        if analysis_type == AnalysisType.WHOLE_GENOME_SEQUENCING:
-            target_bed = self.get_target_bed_from_lims(case_id=case_id) or DEFAULT_CAPTURE_KIT
-        else:
-            target_bed = self.get_target_bed_from_lims(case_id=case_id)
-            if not target_bed:
-                raise ValueError("No capture kit was found in LIMS")
-        return target_bed
-
     def get_workflow_parameters(self, case_id: str) -> RarediseaseParameters:
         """Return parameters."""
         analysis_type: AnalysisType = self.get_analysis_type(case_id=case_id)
-        target_bed: str = self.set_target_bed(case_id=case_id, analysis_type=analysis_type)
+        target_bed: str = self.get_target_bed(case_id=case_id, analysis_type=analysis_type)
         return RarediseaseParameters(
             input=self.get_sample_sheet_path(case_id=case_id),
             outdir=self.get_case_path(case_id=case_id),
